Upgrade to Eclipse Collections 11.0.0.M2 and rename JDK8 classes to JavaStreams.
