+++ deck-of-cards-kata-solutions/pom.xml
@@ -38,7 +38,7 @@
         <maven.compiler.target>16</maven.compiler.target>
 
         <apache-collections.version>4.4</apache-collections.version>
-        <eclipse-collections.version>10.4.0</eclipse-collections.version>
+        <eclipse-collections.version>11.0.0.M2</eclipse-collections.version>
         <guava.version>30.0-jre</guava.version>
         <vavr.version>0.10.3</vavr.version>
         <junit5.version>5.7.0</junit5.version>
