+++ deck-of-cards-kata-solutions/src/main/java/bnymellon/codekatas/deckofcards/sortedset/immutable/JavaStreamsDeckOfCardsAsSortedSet.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package bnymellon.codekatas.deckofcards.list.immutable;
+package bnymellon.codekatas.deckofcards.sortedset.immutable;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
@@ -25,6 +25,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Random;
 import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -32,64 +34,73 @@ import bnymellon.codekatas.deckofcards.Card;
 import bnymellon.codekatas.deckofcards.Rank;
 import bnymellon.codekatas.deckofcards.Suit;
 
-public class JDK8DeckOfCardsAsList
+public class JavaStreamsDeckOfCardsAsSortedSet
 {
-    private final List<Card> cards;
-    private final Map<Suit, List<Card>> cardsBySuit;
+    private final SortedSet<Card> cards;
+    private final Map<Suit, SortedSet<Card>> cardsBySuit;
 
-    public JDK8DeckOfCardsAsList()
+    public JavaStreamsDeckOfCardsAsSortedSet()
     {
-        this.cards = Card.streamCards().sorted().toList();
-        this.cardsBySuit = Map.copyOf(this.cards.stream().collect(Collectors.groupingBy(Card::suit, Collectors.toUnmodifiableList())));
+        this.cards = Collections.unmodifiableSortedSet(
+                Card.streamCards().sorted().collect(Collectors.toCollection(TreeSet::new)));
+        this.cardsBySuit = Map.copyOf(this.cards.stream()
+                .collect(Collectors.groupingBy(
+                        Card::suit,
+                        Collectors.collectingAndThen(
+                                Collectors.toCollection(TreeSet::new),
+                                Collections::unmodifiableSortedSet))));
     }
 
     public Deque<Card> shuffle(Random random)
     {
-        ArrayList<Card> shuffle = new ArrayList<>(this.cards);
-        Collections.shuffle(shuffle, random);
-        Collections.shuffle(shuffle, random);
-        Collections.shuffle(shuffle, random);
+        ArrayList<Card> shuffled = new ArrayList<>(this.cards);
+        Collections.shuffle(shuffled, random);
+        Collections.shuffle(shuffled, random);
+        Collections.shuffle(shuffled, random);
         ArrayDeque<Card> deque = new ArrayDeque<>();
-        shuffle.forEach(deque::push);
+        shuffled.forEach(deque::push);
         return deque;
     }
 
     public Set<Card> deal(Deque<Card> deque, int count)
     {
-        var hand = new HashSet<Card>();
+        Set<Card> hand = new HashSet<>();
         IntStream.range(0, count).forEach(i -> hand.add(deque.pop()));
         return hand;
     }
 
     public List<Set<Card>> shuffleAndDeal(Random random, int hands, int cardsPerHand)
     {
-        var shuffled = this.shuffle(random);
+        Deque<Card> shuffled = this.shuffle(random);
         return this.dealHands(shuffled, hands, cardsPerHand);
     }
 
-    public List<Set<Card>> dealHands(Deque<Card> shuffled, int hands, int cardsPerHand)
+    public List<Set<Card>> dealHands(
+            Deque<Card> shuffled,
+            int hands,
+            int cardsPerHand)
     {
         return IntStream.rangeClosed(1, hands)
                 .mapToObj(i -> this.deal(shuffled, cardsPerHand))
                 .toList();
     }
 
-    public List<Card> diamonds()
+    public SortedSet<Card> diamonds()
     {
         return this.cardsBySuit.get(Suit.DIAMONDS);
     }
 
-    public List<Card> hearts()
+    public SortedSet<Card> hearts()
     {
         return this.cardsBySuit.get(Suit.HEARTS);
     }
 
-    public List<Card> spades()
+    public SortedSet<Card> spades()
     {
         return this.cardsBySuit.get(Suit.SPADES);
     }
 
-    public List<Card> clubs()
+    public SortedSet<Card> clubs()
     {
         return this.cardsBySuit.get(Suit.CLUBS);
     }
@@ -106,12 +117,12 @@ public class JDK8DeckOfCardsAsList
                 .collect(Collectors.groupingBy(Card::rank, Collectors.counting()));
     }
 
-    public List<Card> getCards()
+    public SortedSet<Card> getCards()
     {
         return this.cards;
     }
 
-    public Map<Suit, List<Card>> getCardsBySuit()
+    public Map<Suit, SortedSet<Card>> getCardsBySuit()
     {
         return this.cardsBySuit;
     }
