+++ deck-of-cards-kata-solutions/src/test/java/bnymellon/codekatas/deckofcards/list/immutable/JavaStreamsDeckOfCardsAsListTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Bank of New York Mellon.
+ * Copyright 2021 The Bank of New York Mellon.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,19 @@
  * limitations under the License.
  */
 
-package bnymellon.codekatas.deckofcards.sortedset.immutable;
+package bnymellon.codekatas.deckofcards.list.immutable;
 
-import java.util.Deque;
-import java.util.List;
-import java.util.Map;
 import java.util.Random;
-import java.util.Set;
-import java.util.SortedSet;
 
-import bnymellon.codekatas.deckofcards.Card;
 import bnymellon.codekatas.deckofcards.Rank;
 import bnymellon.codekatas.deckofcards.Suit;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class JDK8DeckOfCardsAsSortedSetTest
+public class JavaStreamsDeckOfCardsAsListTest
 {
-    private JDKImperativeDeckOfCardsAsSortedSet jdk1Deck = new JDKImperativeDeckOfCardsAsSortedSet();
-    private JDK8DeckOfCardsAsSortedSet jdk2Deck = new JDK8DeckOfCardsAsSortedSet();
+    private JDKImperativeDeckOfCardsAsList jdk1Deck = new JDKImperativeDeckOfCardsAsList();
+    private JavaStreamsDeckOfCardsAsList jdk2Deck = new JavaStreamsDeckOfCardsAsList();
 
     @Test
     public void allCards()
@@ -46,7 +40,7 @@ public class JDK8DeckOfCardsAsSortedSetTest
         var jdk2Cards = this.jdk2Deck.getCards();
         Assertions.assertThrows(
                 UnsupportedOperationException.class,
-                () -> jdk2Cards.remove(null));
+                () -> jdk2Cards.remove(0));
         Assertions.assertThrows(
                 UnsupportedOperationException.class,
                 jdk2Cards::clear);
@@ -82,37 +76,37 @@ public class JDK8DeckOfCardsAsSortedSetTest
     @Test
     public void deal()
     {
-        Deque<Card> jdk1Shuffle = this.jdk1Deck.shuffle(new Random(1));
-        Deque<Card> jdk2Shuffle = this.jdk2Deck.shuffle(new Random(1));
+        var jdk1Shuffle = this.jdk1Deck.shuffle(new Random(1));
+        var jdk2Shuffle = this.jdk2Deck.shuffle(new Random(1));
 
-        Set<Card> jdk1Hand = this.jdk1Deck.deal(jdk1Shuffle, 5);
-        Set<Card> jdk2Hand = this.jdk2Deck.deal(jdk2Shuffle, 5);
+        var jdk1Hand = this.jdk1Deck.deal(jdk1Shuffle, 5);
+        var jdk2Hand = this.jdk2Deck.deal(jdk2Shuffle, 5);
         Assertions.assertEquals(jdk1Hand, jdk2Hand);
     }
 
     @Test
     public void shuffleAndDealHands()
     {
-        List<Set<Card>> jdk1Hands = this.jdk1Deck.shuffleAndDeal(new Random(1), 5, 5);
-        List<Set<Card>> jdk2Hands = this.jdk2Deck.shuffleAndDeal(new Random(1), 5, 5);
+        var jdk1Hands = this.jdk1Deck.shuffleAndDeal(new Random(1), 5, 5);
+        var jdk2Hands = this.jdk2Deck.shuffleAndDeal(new Random(1), 5, 5);
         Assertions.assertEquals(jdk1Hands, jdk2Hands);
     }
 
     @Test
     public void dealHands()
     {
-        Deque<Card> jdk1Shuffled = this.jdk1Deck.shuffle(new Random(1));
-        Deque<Card> jdk2Shuffled = this.jdk2Deck.shuffle(new Random(1));
-        List<Set<Card>> jdk1Hands = this.jdk1Deck.dealHands(jdk1Shuffled, 5, 5);
-        List<Set<Card>> jdk2Hands = this.jdk2Deck.dealHands(jdk2Shuffled, 5, 5);
+        var jdk1Shuffled = this.jdk1Deck.shuffle(new Random(1));
+        var jdk2Shuffled = this.jdk2Deck.shuffle(new Random(1));
+        var jdk1Hands = this.jdk1Deck.dealHands(jdk1Shuffled, 5, 5);
+        var jdk2Hands = this.jdk2Deck.dealHands(jdk2Shuffled, 5, 5);
         Assertions.assertEquals(jdk1Hands, jdk2Hands);
     }
 
     @Test
     public void cardsBySuit()
     {
-        Map<Suit, SortedSet<Card>> jdk1CardsBySuit = this.jdk1Deck.getCardsBySuit();
-        Map<Suit, SortedSet<Card>> jdk2CardsBySuit = this.jdk2Deck.getCardsBySuit();
+        var jdk1CardsBySuit = this.jdk1Deck.getCardsBySuit();
+        var jdk2CardsBySuit = this.jdk2Deck.getCardsBySuit();
         Assertions.assertEquals(jdk1CardsBySuit.get(Suit.CLUBS), jdk2CardsBySuit.get(Suit.CLUBS));
     }
 
@@ -128,7 +122,7 @@ public class JDK8DeckOfCardsAsSortedSetTest
                 jdk2CardsBySuit::clear);
         Assertions.assertThrows(
                 UnsupportedOperationException.class,
-                () -> jdk2CardsBySuit.get(Suit.CLUBS).remove(null));
+                () -> jdk2CardsBySuit.get(Suit.CLUBS).remove(0));
         Assertions.assertThrows(
                 UnsupportedOperationException.class,
                 () -> jdk2CardsBySuit.get(Suit.CLUBS).add(null));
