+++ deck-of-cards-kata/src/main/java/bnymellon/codekatas/deckofcards/sortedset/immutable/JavaStreamsDeckOfCardsAsSortedSet.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package bnymellon.codekatas.deckofcards.list.immutable;
+package bnymellon.codekatas.deckofcards.sortedset.immutable;
 
 import java.util.Collections;
 import java.util.Deque;
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Random;
 import java.util.Set;
+import java.util.SortedSet;
 import java.util.function.Function;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;
@@ -32,20 +33,20 @@ import bnymellon.codekatas.deckofcards.Card;
 import bnymellon.codekatas.deckofcards.Rank;
 import bnymellon.codekatas.deckofcards.Suit;
 
-public class JDK8DeckOfCardsAsList
+public class JavaStreamsDeckOfCardsAsSortedSet
 {
-    private List<Card> cards;
-    private Map<Suit, List<Card>> cardsBySuit;
+    private SortedSet<Card> cards;
+    private Map<Suit, SortedSet<Card>> cardsBySuit;
 
     /**
      * TODO Replace the null values below with something more useful.
-     * Use Java 8 Streams API with {@link Card#streamCards()} to create an "immutable" List and store that in cards.
+     * Use Java 8 Streams APIs with {@link Card#streamCards()} to create an "immutable" SortedSet and store that in cards.
      * Group all of the cards by {@link Card#suit()} into an "immutable" Map and stored that in cardsBySuit.
      *
-     * Hint: Look at {@link Collectors#groupingBy(Function)}, {@link Collections#unmodifiableList(List)}
+     * Hint: Look at {@link Collectors#groupingBy(Function)}, {@link Collections#unmodifiableSortedSet(SortedSet)}
      * {@link Collectors#collectingAndThen(Collector, Function)}, {@link Collectors#mapping(Function, Collector)}
      */
-    public JDK8DeckOfCardsAsList()
+    public JavaStreamsDeckOfCardsAsSortedSet()
     {
         this.cards = null;
         this.cardsBySuit = null;
@@ -54,50 +55,53 @@ public class JDK8DeckOfCardsAsList
     public Deque<Card> shuffle(Random random)
     {
         // TODO Shuffle the deck 3 times with the Random parameter and push the shuffled cards onto an ArrayDeque
-        // Hint: Look at IntStream.range() or IntStream.rangeClosed() and Collections.shuffle().
+        // Hint: Look at IntStream.range() or IntStream.rangeClosed() and Collections.shuffle()
         return null;
     }
 
     public Set<Card> deal(Deque<Card> deque, int count)
     {
-        var hand = new HashSet<Card>();
+        Set<Card> hand = new HashSet<>();
         IntStream.range(0, count).forEach(i -> hand.add(deque.pop()));
         return hand;
     }
 
     public List<Set<Card>> shuffleAndDeal(Random random, int hands, int cardsPerHand)
     {
-        var shuffled = this.shuffle(random);
+        Deque<Card> shuffled = this.shuffle(random);
         return this.dealHands(shuffled, hands, cardsPerHand);
     }
 
-    public List<Set<Card>> dealHands(Deque<Card> shuffled, int hands, int cardsPerHand)
+    public List<Set<Card>> dealHands(
+            Deque<Card> shuffled,
+            int hands,
+            int cardsPerHand)
     {
         // TODO Deal the number of hands with the cardsPerHand into an "immutable" List<Set<Card>>
         return null;
     }
 
-    public List<Card> diamonds()
+    public SortedSet<Card> diamonds()
     {
-        // TODO return all diamonds as an "Immutable" List
+        // TODO return all diamonds as an "Immutable" SortedSet
         return null;
     }
 
-    public List<Card> hearts()
+    public SortedSet<Card> hearts()
     {
-        // TODO return all hearts as an "Immutable" List
+        // TODO return all hearts as an "Immutable" SortedSet
         return null;
     }
 
-    public List<Card> spades()
+    public SortedSet<Card> spades()
     {
-        // TODO return all spades as an "Immutable" List
+        // TODO return all spades as an "Immutable" SortedSet
         return null;
     }
 
-    public List<Card> clubs()
+    public SortedSet<Card> clubs()
     {
-        // TODO return all clubs as an "Immutable" List
+        // TODO return all clubs as an "Immutable" SortedSet
         return null;
     }
 
@@ -115,12 +119,12 @@ public class JDK8DeckOfCardsAsList
         return null;
     }
 
-    public List<Card> getCards()
+    public SortedSet<Card> getCards()
     {
         return this.cards;
     }
 
-    public Map<Suit, List<Card>> getCardsBySuit()
+    public Map<Suit, SortedSet<Card>> getCardsBySuit()
     {
         return this.cardsBySuit;
     }
