+++ java-lambda-kata/pom.xml
@@ -37,7 +37,7 @@
         <maven.compiler.source>16</maven.compiler.source>
         <maven.compiler.target>16</maven.compiler.target>
 
-        <eclipse-collections.version>10.4.0</eclipse-collections.version>
+        <eclipse-collections.version>11.0.0.M2</eclipse-collections.version>
         <junit5.version>5.7.0</junit5.version>
     </properties>
